backend/requirements.txt
flask==2.2.5
flask-sqlalchemy==3.1.1
flask-migrate==4.1.0
python-dotenv==0.21.0
mysqlclient==2.2.7
